requirements.txt
Flask==3.0.3
PyMySQL==1.1.0
Flask-SQLAlchemy==3.1.1
PyYAML==6.0.1
flasgger==0.9.7.1
python-dotenv==1.0.0
gunicorn==21.2.0
